+++ tests_react_native/__tests__/validators/validateIOSNotification.test.ts
@@ -1,8 +1,8 @@
 import validateIOSNotification from '@notifee/react-native/src/validators/validateIOSNotification';
 import { NotificationIOS } from '@notifee/react-native/src/types/NotificationIOS';
 
-describe('Validate IOS Input', () => {
-  describe('validateIOSInput()', () => {
+describe('Validate IOS Notification', () => {
+  describe('validateIOSNotification()', () => {
     test('returns valid ', () => {
       const notification: NotificationIOS = {
         attachments: [],
@@ -56,6 +56,34 @@ describe('Validate IOS Input', () => {
         },
       });
       expect($).toEqual({
+        foregroundPresentationOptions: {
+          alert: true,
+          badge: true,
+          sound: true,
+          banner: true,
+          list: true,
+        },
+      });
+    });
+
+    test('returns valid when there is a valid communicationInfo property', () => {
+      const $ = validateIOSNotification({
+        communicationInfo: {
+          conversationId: 'id',
+          sender: {
+            id: 'sender-id',
+            displayName: 'John Doe',
+          },
+        },
+      });
+      expect($).toEqual({
+        communicationInfo: {
+          conversationId: 'id',
+          sender: {
+            id: 'sender-id',
+            displayName: 'John Doe',
+          },
+        },
         foregroundPresentationOptions: {
           alert: true,
           badge: true,
@@ -151,6 +179,24 @@ describe('Validate IOS Input', () => {
 
       expect(() => validateIOSNotification(notification)).toThrowError(
         "'notification.ios.launchImageName' expected a string value.",
+      );
+    });
+
+    test('returns invalid when an invalid communicationInfo property is provided', () => {
+      let notification: NotificationIOS = {
+        communicationInfo: '' as any,
+      };
+
+      expect(() => validateIOSNotification(notification)).toThrowError(
+        "'ios.communicationInfo' expected an object.",
+      );
+
+      notification = {
+        communicationInfo: {} as any,
+      };
+
+      expect(() => validateIOSNotification(notification)).toThrowError(
+        "'ios.communicationInfo' 'conversationId' expected a valid string value.",
       );
     });
   });