+++ tests_react_native/example/notifications.ts
@@ -4,8 +4,8 @@ import {
   AndroidLaunchActivityFlag,
   AndroidCategory,
   AndroidImportance,
+  AndroidFlags,
 } from '@notifee/react-native';
-import { AndroidFlags } from '@notifee/react-native/src';
 
 export const notifications: { key: string; notification: Notification | Notification[] }[] = [
   {
@@ -51,6 +51,30 @@ export const notifications: { key: string; notification: Notification | Notifica
         },
       },
       ios: {},
+    },
+  },
+  {
+    key: 'Communications',
+    notification: {
+      id: 'communication-id',
+      title: 'Communication PN',
+      android: {
+        channelId: 'high',
+        pressAction: {
+          id: 'default',
+        },
+      },
+      ios: {
+        categoryId: 'communicationId',
+        communicationInfo: {
+          conversationId: '123',
+          sender: {
+            id: 'abcde',
+            avatar: 'https://pbs.twimg.com/profile_images/1070077650713133056/oji2RT4i_normal.jpg',
+            displayName: 'Helena Ford',
+          },
+        },
+      },
     },
   },
   {